[ADD] Call Modal + some fixes

+++ client/src/components/message/RightSide.js
@@ -141,9 +141,34 @@ const RightSide = () => {
     
         
         const handleDeleteConversation = () => {
-            dispatch(deleteConversation({auth, id}))
-            return history.push('/message')
+            if (window.confirm('Do you want to delete this conversation?')) {
+                dispatch(deleteConversation({auth, id}))
+                return history.push('/message') 
+            }
         }
+
+        // Call
+        const caller = ({video}) => {
+            const { _id, avatar, username, fullname } = user
+
+            const msg = {
+                sender: auth.user._id,
+                recipient: _id,
+                avatar, username, fullname, video
+            }
+            dispatch({ type: GLOBALTYPES.CALL, payload: msg })
+        }
+
+        
+        const handleAudioCall = () => {
+            caller({video: false})
+        }
+
+        
+        const handleVideoCall = () => {
+            caller({video: true})
+        }
+
 
 
     return (
@@ -152,8 +177,16 @@ const RightSide = () => {
                 {
                     user.length !== 0 &&
                     <UserCard user={user}>
-                        <i className="fas fa-trash text-danger" 
-                        onClick={handleDeleteConversation}/>
+                        <div>
+                            <i className="fas fa-phone-alt" 
+                            onClick={handleAudioCall}/>
+
+                            <i className="fas fa-video mx-3" 
+                            onClick={handleVideoCall}/>
+
+                            <i className="fas fa-trash text-danger" 
+                            onClick={handleDeleteConversation}/>
+                        </div>
                     </UserCard>
                 }
             </div>
